split controller
