+++ resources/views/layouts/admin/sidebar.blade.php
@@ -75,7 +75,7 @@
                     </a>
                 </li>
                 <li class="nav-item has-submenu">
-                    <a class="tw-cursor-pointer nav-link">
+                    <a class="tw-cursor-pointer nav-link" href="{{ route('laporan') }}"">
                         <i class="nav-icon fa fa-file tw-text-prim-white"></i>
                         <p class="tw-text-prim-red tw-font-bold focus:tw-text-prim-white">
                             Laporan
