+++ resources/views/layouts/master/supplier.blade.php
@@ -125,9 +125,7 @@
                     'X-CSRF-TOKEN': '{{csrf_token()}}'
                 }
             },
-            language: {
-                processing: '<i class="fa fa-spinner fa-spin"></i> Tunggu Sebentar'
-            },
+        
             columns: [{
                     data: 'DT_RowIndex',
                     className: 'nowrap-text align-center',
